use refs for blocks
