+++ packages/react/src/ReactInlineContentSpec.tsx
@@ -15,8 +15,7 @@ import {
 } from "@tiptap/react";
 // import { useReactNodeView } from "@tiptap/react/dist/packages/react/src/useReactNodeView";
 import { FC } from "react";
-import { flushSync } from "react-dom";
-import { createRoot } from "react-dom/client";
+import { renderToDOMSpec } from "./ReactRenderUtil";
 
 // this file is mostly analogoues to `customBlocks.ts`, but for React blocks
 
@@ -71,44 +70,11 @@ export function createReactInlineContentSpec<
         editor.inlineContentSchema,
         editor.styleSchema
       ) as any as InlineContentFromConfig<T, S>; // TODO: fix cast
-
       const Content = inlineContentImplementation.render;
 
-      let contentDOM: HTMLElement | undefined;
-      const div = document.createElement("div");
-      const root = createRoot(div);
-      flushSync(() => {
-        root.render(
-          <Content
-            inlineContent={ic}
-            contentRef={(el) => (contentDOM = el || undefined)}
-          />
-        );
-      });
-
-      if (!div.childElementCount) {
-        // TODO
-        console.warn("ReactInlineContentSpec: renderHTML() failed");
-        return {
-          dom: document.createElement("span"),
-        };
-      }
-
-      // clone so we can unmount the react root
-      contentDOM?.setAttribute("data-tmp-find", "true");
-      const cloneRoot = div.cloneNode(true) as HTMLElement;
-      const dom = cloneRoot.firstElementChild! as HTMLElement;
-      const contentDOMClone = cloneRoot.querySelector(
-        "[data-tmp-find]"
-      ) as HTMLElement | null;
-      contentDOMClone?.removeAttribute("data-tmp-find");
-
-      root.unmount();
-
-      return {
-        dom,
-        contentDOM: contentDOMClone || undefined,
-      };
+      return renderToDOMSpec((refCB) => (
+        <Content inlineContent={ic} contentRef={refCB} />
+      ));
     },
 
     // TODO: needed?
